requirements.txt
streamlit>=1.36.0
pandas>=2.0.0
openpyxl>=3.1.0
google-generativeai>=0.7.0
